(develop) Separate routes out and organize routing and entry point (index.js)
